Starts to modularise the game code and split it out into separate files for better maintainability
