+++ src/layouts/ResumeAdventureLayout.astro
@@ -52,8 +52,9 @@ _______ ______          _______ _      _________        _______ _______
 						<div class="loading-text">
 							<div class="loading"></div>
 							<p><small>(Apologies to Lord British)</small></p>
-							<p>ESC - Just give me a CV</p>
-							<p>RETURN - Play along</p>
+							<p id="escape">ESC - Just give me a CV</p>
+							<p id="play">RETURN - Play along</p>
+                            <p>(Mobile users: tap an option above if you don't have a keyboard)</p>
 						</div>
 					</div>
 				</div>
@@ -89,6 +90,6 @@ _______ ______          _______ _      _________        _______ _______
 				</div>
 			</div>
 		</main>
-		<script src="../scripts/portfolio/portfolio.js"></script>
+		<script src="../scripts/portfolio.js"></script>
 	</body>
 </html>
